+++ examples/next-openai/app/api/use-chat-persistence-single-message/route.ts
@@ -3,9 +3,10 @@ import { loadChat, saveChat } from '@util/chat-store';
 import { convertToModelMessages, streamText, UIMessage } from 'ai';
 
 export async function POST(req: Request) {
-  const { message, id }: { message: UIMessage; id: string } = await req.json();
+  const { message, chatId }: { message: UIMessage; chatId: string } =
+    await req.json();
 
-  const previousMessages = await loadChat(id);
+  const previousMessages = await loadChat(chatId);
   const messages = [...previousMessages, message];
 
   const result = streamText({
@@ -16,7 +17,7 @@ export async function POST(req: Request) {
   return result.toUIMessageStreamResponse({
     originalMessages: messages,
     onFinish: ({ messages }) => {
-      saveChat({ id, messages });
+      saveChat({ chatId, messages });
     },
   });
 }