+++ packages/ai/src/ui-message-stream/ui-message-stream-parts.ts
@@ -81,6 +81,12 @@ export const uiMessageStreamPartSchema = z.union([
   }),
 ]);
 
+export type DataUIMessageStreamPart = {
+  type: `data-${string}`;
+  id?: string;
+  data: unknown;
+};
+
 export type UIMessageStreamPart =
   | {
       type: 'text';
@@ -131,11 +137,7 @@ export type UIMessageStreamPart =
       url: string;
       mediaType: string;
     }
-  | {
-      type: `data-${string}`;
-      id?: string;
-      data: unknown;
-    }
+  | DataUIMessageStreamPart
   | {
       type: 'metadata';
       metadata: unknown;
@@ -160,3 +162,9 @@ export type UIMessageStreamPart =
   | {
       type: 'reasoning-part-finish';
     };
+
+export function isDataUIMessageStreamPart(
+  part: UIMessageStreamPart,
+): part is DataUIMessageStreamPart {
+  return part.type.startsWith('data-');
+}