+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/vue
 
+## 2.0.0-alpha.1
+
+### Patch Changes
+
+- Updated dependencies [b346545]
+- Updated dependencies [109c0ac]
+  - ai@5.0.0-alpha.1
+  - @ai-sdk/provider-utils@3.0.0-alpha.1
+
 ## 2.0.0-canary.23
 
 ### Patch Changes